pull dynarec out into its own module
