+++ pchan-emu/src/cpu/ops/mult.rs
@@ -101,7 +101,7 @@ mod tests {
         #[case] b: u32,
         #[case] expected: u64,
     ) -> color_eyre::Result<()> {
-        use crate::JitSummary;
+        use crate::dynarec::JitSummary;
 
         emulator
             .mem
@@ -130,7 +130,7 @@ mod tests {
     ) -> color_eyre::Result<()> {
         assert!(a == 0 || b == 0);
 
-        use crate::JitSummary;
+        use crate::dynarec::JitSummary;
 
         emulator
             .mem
