+++ pchan-emu/src/cpu/ops/or.rs
@@ -89,7 +89,7 @@ mod tests {
     use pretty_assertions::assert_eq;
     use rstest::rstest;
 
-    use crate::JitSummary;
+    use crate::dynarec::JitSummary;
     use crate::cpu::ops::prelude::*;
     use crate::{Emu, memory::KSEG0Addr, test_utils::emulator};
