+++ pchan-emu/src/cpu/ops/srl.rs
@@ -116,7 +116,7 @@ mod tests {
         #[case] b: i16,
         #[case] expected: u32,
     ) -> color_eyre::Result<()> {
-        use crate::JitSummary;
+        use crate::dynarec::JitSummary;
 
         emulator.mem.write_array(
             KSEG0Addr::from_phys(0),
@@ -131,7 +131,7 @@ mod tests {
     #[case(8)]
     #[case(0b00001111)]
     fn srl_2(setup_tracing: (), mut emulator: Emu, #[case] imm: i16) -> color_eyre::Result<()> {
-        use crate::JitSummary;
+        use crate::dynarec::JitSummary;
 
         emulator
             .mem
